Move log out button to navbar
