+++ src/@commitUI/components/Input/Input.module.css
@@ -14,7 +14,7 @@
   border-radius: 8px;
   border: var(--gray-lines) 1px solid;
   padding: 24px 16px 4px 16px;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
   font-size: 16px;
   line-height: 1;
   outline: none;
