+++ src/components/Navbar/Navbar.module.scss
@@ -1,30 +1,13 @@
-.hamburger {
-    display: none;
-    color: #fff;
-    padding: 0;
-    margin: 0;
-    margin-right: 16px;
-}
-
-.navmenu {
+.nav {
     display: flex;
-    align-items: center;
+    justify-content: space-between;
 }
 
-.btn {
-    margin: var(--spacing-sm);
+.align {
+    display: flex;
 }
 
-@media screen and (max-width: 768px) {
-    .hamburger {
-        display: block;
-        z-index: 1001;
-        top: 0;
-        right: 0;
-        cursor: pointer;
-    }
-
-    .navmenu {
-        display: none;
-    }
+.button {
+    width: 200px;
+    margin-right: 16px;
 }
